Exercise2

Description

Predefined Types

data UNIT

Unit

Constructors

UNIT

data EITHER a b

Sum

Constructors

LEFT a

RIGHT b

data PAIR a b

Product

Constructors

PAIR a b

data CONS a

Constructor

Constructors

CONS String a

data Bin a

Binary trees

Constructors

Leaf

Bin (Bin a) a (Bin a)

type BinG a = EITHER (CONS UNIT) (CONS (PAIR (Bin a) (PAIR a (Bin a))))

Generic representation of a Binary tree

type ListG a = EITHER (CONS UNIT) (CONS (PAIR a [a]))

Generic representation of a List

Review Questions

1) We have three implementations of (==) for UNIT. All three give the same results because UNIT only has one state. As for the code, I like third implementation the most.

2) The String parameter in CONS constructor should, in principle, by a function of type a. So there's no need to compare these names -- they are the same by construction.

3) Generic representations of both Leaf and [] is UNIT. Leaf == [] gives compile-time error :) No matter how you define it. Type of (==) is a → a, so comparing Bin with List is illformed.

Generic Serialisation

There's a design problem with our definition of CONS. It tries to keep the info about constructor names storing them in a String field. This info, however, in only available at runtime. So if we read a CONS UNIT, there's no way to verify that constructor name we read actually makes sense. So CONS is OK with both

["CONS", "[]", "UNIT"]

and

["CONS", "Abrakadabra", "UNIT"]

So why is this a problem? According to the exercise, toList function doesn't fail! This means that LEFT (CONS "Abrakadabra" UNIT) must produce a valid list. OK, suppose we neglect the name parameter and just construct an empty list. Then we'll have problems when no generic information is provided. Consider

["[]"]

vs.

["(:)", "123", "(", "[]", ")"]

We can distinguish these two cases by looking at number of arguments, right? But what if out type is

data Foo = A Int Int | B Int Int

Now we really need to take the constructor name into account. This, however, is impossible when reading a CONS object, because constructor name is not part of the type and is hence unknown at compile-time. We thus can only trigger an error when converting to List. So it should be

toList ∷ ListG a → Maybe [a]

This, however, is not enough. We need to keep track of all possible ways of parsing input. So our generic read function should be

read' ∷ [String] → [(a, [String])]

And because Haskell has support for ReadS, let's make it

read' ∷ String → [(a, String)]

With these modifications we can now implement everything :)

P.S. I think this also answers the "Reflection" questions.

fromList :: [a] -> ListG a

Converts a list to its generic representation. This function is inverse of toList. Constructor for unit list is called "[]", for composite list -- "(:)".

toList :: ListG a -> Maybe [a]

Converts a generic representation of a list to an actual list. This function is inverse of fromList. It may fail if constructors have wrong names.

fromBin :: Bin a -> BinG a

Converts a binary tree to its generic representaion. This function is inverse of toBin. Constructor for unit tree is called Leaf, for composite tree -- Bin.

toBin :: BinG a -> Maybe (Bin a)

Converts a generic representation of a binary tree back to an actual binary tree. This function is inverse of fromBin. It may fail if constructors have wrong names

With Generic Information

class VSerialise α where

Verbose version of Serialise typeclass, i.e. generic type constructors are included in the serialised data.

Minimal complete definition

vreadPrec', vwritePrec'

Methods

vreadPrec' :: Int -> ReadS α

vwritePrec' :: Int -> α -> String

vread' :: VSerialise α => String -> Maybe α

Our hand-made verbose analogue of read.

vwrite' :: VSerialise α => α -> String

Our hand-made verbose analogue of show.

Without Generic Information

class Serialise α where

Compact Serialise typeclass, i.e. no generic information is included in the serialised data.

Minimal complete definition

readPrec', writePrec'

Methods

readPrec' :: Int -> ReadS α

writePrec' :: Int -> α -> String

read' :: Serialise α => String -> Maybe α

Our hand-made analogue of show.

write' :: Serialise α => α -> String

Our hand-made analogue of show.

test :: IO ()

Main